CreativeScrapyard/requirements.txt
django-mathfilters
pycryptodome
django-filter
xhtml2pdf

# WeasyPrint
# django-report-builder